+++ acceptance-tests/test-plugins/src/main/java/org/hyperledger/besu/tests/acceptance/plugins/TestPermissioningPlugin.java
@@ -19,6 +19,7 @@ import org.hyperledger.besu.plugin.ServiceManager;
 import org.hyperledger.besu.plugin.services.PermissioningService;
 import org.hyperledger.besu.plugin.services.PicoCLIOptions;
 
+
 import com.google.auto.service.AutoService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -75,6 +76,17 @@ public class TestPermissioningPlugin implements BesuPlugin {
             }
             return true;
           });
+
+      service.registerTransactionPermissioningProvider(
+              transaction -> {
+                long configuredGasLimitThreshold = 12000L;
+                long gasLimit = transaction.getGasLimit();
+                LOG.info("Transaction gas limit: {} | Configured threshold: {} ", gasLimit, configuredGasLimitThreshold);
+                return gasLimit > configuredGasLimitThreshold;
+              }
+      );
+
+
     }
   }